requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.27
aiogram==3.4.1
requests~=2.31.0
tronpy
psycopg2